readme: update helm install command
With the previous command helm would fail with the following error:
```bash
$ helm install nats-nui/nui
Error: INSTALLATION FAILED: must either provide a name or specify --generate-name
```

For the lazy people like me having a one liner that just works would be
great. Hence with this I'm proposing to give it a default name as
`nats-nui` to solve the above error.